+++ ving/schema/validator.mjs
@@ -85,13 +85,13 @@ export const validateProps = (schema) => {
 export const validatePropRelation = (prop, schema) => {
     if (!['id', 'virtual'].includes(prop.type)) {
         if ('relation' in prop)
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation should not exist.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation should not exist.`);
         return;
     }
     if (prop.type == 'id' && !('relation' in prop)) // not required
         return;
     if (!isObject(prop.relation))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation must be an object.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation must be an object.`);
     validateRelationType(prop, schema);
     validateRelationKind(prop, schema);
     validateRelationName(prop, schema);
@@ -112,19 +112,19 @@ export const validateRelationAcceptedFileExtensions = (prop, schema) => {
         || !['parent'].includes(prop.relation.type) // excluded by type
     ) {
         if ('acceptedFileExtensions' in prop.relation)
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation.acceptedFileExtensions should not exist.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation.acceptedFileExtensions should not exist.`);
         return;
     }
     if (!('acceptedFileExtensions' in prop.relation)) // optional
         return;
     if (!isArray(prop.relation.acceptedFileExtensions))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation.acceptedFileExtensions must be an array.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation.acceptedFileExtensions must be an array.`);
     if (prop.relation.acceptedFileExtensions.length == 0)
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation.acceptedFileExtensions must have at least 1 file extension.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation.acceptedFileExtensions must have at least 1 file extension.`);
     const allowedExtensions = Object.keys(extensionMap);
     for (const ext of prop.relation.acceptedFileExtensions) {
         if (!allowedExtensions.includes(ext))
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation.acceptedFileExtensions value ${ext} is not an extension listed in the S3File extensionMap.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation.acceptedFileExtensions value ${ext} is not an extension listed in the S3File extensionMap.`);
     }
 }
 
@@ -139,13 +139,13 @@ export const validateRelationAcceptedFileExtensions = (prop, schema) => {
 export const validateRelationSkipOwnerCheck = (prop, schema) => {
     if (!['parent'].includes(prop.relation.type)) {// excluded by type
         if ('skipOwnerCheck' in prop.relation)
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation.skipOwnerCheck should not exist.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation.skipOwnerCheck should not exist.`);
         return;
     }
     if (!('skipOwnerCheck' in prop.relation)) // optional
         return;
     if (!isBoolean(prop.relation.skipOwnerCheck))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation.skipOwnerCheck must be a boolean.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation.skipOwnerCheck must be a boolean.`);
 }
 
 /**
@@ -158,9 +158,9 @@ export const validateRelationSkipOwnerCheck = (prop, schema) => {
  */
 export const validateRelationName = (prop, schema) => {
     if (!isString(prop.relation.name))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation.name must be a string.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation.name must be a string.`);
     if (!/^[a-z]/.test(prop.relation.name))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation.name must start with a lower case letter.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation.name must start with a lower case letter.`);
 }
 
 /**
@@ -173,14 +173,14 @@ export const validateRelationName = (prop, schema) => {
  */
 export const validateRelationKind = (prop, schema) => {
     if (!isString(prop.relation.kind))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation.kind must be a string.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation.kind must be a string.`);
     const otherSchema = vingSchemas.find(s => s.kind == prop.relation.kind);
     if (isUndefined(otherSchema))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation.kind refers to a schema that doesn't exist.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation.kind refers to a schema that doesn't exist.`);
     if (prop.relation.type == 'child') {
         const childProp = otherSchema.props.find(p => p.name == prop.name);
         if (isUndefined(childProp))
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name} refers to ${prop.relation.kind}.props.${prop.name} which does not exist.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)} refers to ${prop.relation.kind}.props.${prop.name} which does not exist.`);
     }
 }
 
@@ -194,13 +194,13 @@ export const validateRelationKind = (prop, schema) => {
  */
 export const validateRelationType = (prop, schema) => {
     if (!isString(prop.relation.type))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation.type must be a string.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation.type must be a string.`);
     if (!['child', 'parent'].includes(prop.relation.type))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation.type must either child or parent.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation.type must either child or parent.`);
     if (prop.relation.type == 'parent' && prop.type != 'id')
-        throw ving.ouch(442, `${schema.kind}.props.type must be id when ${schema.kind}.props.${prop.name}.relation.type is parent.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.type must be id when ${formatPropPath(prop, schema)}.relation.type is parent.`);
     if (prop.relation.type == 'child' && prop.type != 'virtual')
-        throw ving.ouch(442, `${schema.kind}.props.type must be virtual when ${schema.kind}.props.${prop.name}.relation.type is child.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.type must be virtual when ${formatPropPath(prop, schema)}.relation.type is child.`);
 }
 
 /**
@@ -214,13 +214,13 @@ export const validateRelationType = (prop, schema) => {
 export const validatePropAutoUpdate = (prop, schema) => {
     if (!['date'].includes(prop.type)) {
         if ('autoUpdate' in prop)
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.autoUpdate should not exist.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.autoUpdate should not exist.`);
         return;
     }
     if (!('autoUpdate' in prop)) // not required
         return;
     if (!isBoolean(prop.autoUpdate))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.autoUpdate must be a boolean.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.autoUpdate must be a boolean.`);
 }
 
 /**
@@ -235,17 +235,17 @@ export const validatePropUniqueQualfiers = (prop, schema) => {
     if (!('uniqueQualifiers' in prop)) // optional
         return
     if (prop.unique != true)
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.uniqueQualifiers should not exist.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.uniqueQualifiers should not exist.`);
     if (!isArray(prop.uniqueQualifiers))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.uniqueQualifiers must be an array.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.uniqueQualifiers must be an array.`);
     if (prop.uniqueQualifiers.length == 0)
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.uniqueQualifiers must have at least one member.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.uniqueQualifiers must have at least one member.`);
     for (const qualifier of prop.uniqueQualifiers) {
         const otherProp = schema.props.find(p => p.name == qualifier);
         if (isUndefined(otherProp))
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.uniqueQualifiers value ${qualifier} does not match another prop name.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.uniqueQualifiers value ${qualifier} does not match another prop name.`);
         if (otherProp.name == prop.name)
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.uniqueQualifiers value ${qualifier} references its own prop.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.uniqueQualifiers value ${qualifier} references its own prop.`);
     }
 }
 
@@ -260,13 +260,13 @@ export const validatePropUniqueQualfiers = (prop, schema) => {
 export const validatePropUnique = (prop, schema) => {
     if (!['string'].includes(prop.type)) {
         if ('unique' in prop)
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.unique should not exist.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.unique should not exist.`);
         return;
     }
     if (!('unique' in prop)) // optional
         return
     if (!isBoolean(prop.unique))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.unique must be a boolean.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.unique must be a boolean.`);
 }
 
 /**
@@ -280,13 +280,13 @@ export const validatePropUnique = (prop, schema) => {
 export const validatePropLength = (prop, schema) => {
     if (!['id', 'string'].includes(prop.type)) {
         if ('length' in prop)
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.length should not exist.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.length should not exist.`);
         return;
     }
     if (!isNumber(prop.length))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.length must be a number.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.length must be a number.`);
     if (prop.length < 1)
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.length must be greater than 0.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.length must be greater than 0.`);
 }
 
 /**
@@ -300,16 +300,16 @@ export const validatePropLength = (prop, schema) => {
 export const validatePropEnumLabels = (prop, schema) => {
     if (!['enum', 'boolean'].includes(prop.type)) {
         if ('enumLabels' in prop)
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.enumLabels should not exist.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.enumLabels should not exist.`);
         return;
     }
     if (!isArray(prop.enumLabels))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.enumLabels must be an array.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.enumLabels must be an array.`);
     if (prop.enums.length != prop.enumLabels.length)
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.enumLabels must have the same number of options as enums.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.enumLabels must have the same number of options as enums.`);
     for (const value of prop.enumLabels) {
         if (!isString(value))
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.enumLabels values must be strings.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.enumLabels values must be strings.`);
     }
 }
 
@@ -324,18 +324,18 @@ export const validatePropEnumLabels = (prop, schema) => {
 export const validatePropEnums = (prop, schema) => {
     if (!['enum', 'boolean'].includes(prop.type)) {
         if ('enums' in prop)
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.enums should not exist.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.enums should not exist.`);
         return;
     }
     if (!isArray(prop.enums))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.enums must be an array.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.enums must be an array.`);
     if (prop.enums.length == 0)
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.enums must have at least one option.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.enums must have at least one option.`);
     for (const value of prop.enums) {
         if (prop.type == 'boolean' && !isBoolean(value))
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.enums values must be boolean.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.enums values must be boolean.`);
         if (prop.type == 'enum' && !isString(value))
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.enums values must be strings.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.enums values must be strings.`);
     }
 }
 
@@ -349,7 +349,7 @@ export const validatePropEnums = (prop, schema) => {
  */
 export const validatePropEdit = (prop, schema) => {
     if (!isArray(prop.edit))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.edit must be an array.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.edit must be an array.`);
     for (const role of prop.edit) {
         if (['public', 'owner'].includes(role)) {
             // we're a special keyword that we know
@@ -358,7 +358,7 @@ export const validatePropEdit = (prop, schema) => {
             // we're a known role
         }
         else {
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.edit role ${role} is unknown.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.edit role ${role} is unknown.`);
         }
     }
 }
@@ -373,7 +373,7 @@ export const validatePropEdit = (prop, schema) => {
  */
 export const validatePropView = (prop, schema) => {
     if (!isArray(prop.view))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.view must be an array.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.view must be an array.`);
     for (const role of prop.view) {
         if (['public', 'owner'].includes(role)) {
             // we're a special keyword that we know
@@ -382,7 +382,7 @@ export const validatePropView = (prop, schema) => {
             // we're a known role
         }
         else {
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.view role ${role} is unknown.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.view role ${role} is unknown.`);
         }
     }
 }
@@ -402,7 +402,7 @@ export const validatePropZod = (prop, schema) => {
     )
         return;
     if (!isFunction(prop.zod))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.zod must be a function that returns a zod validation function.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.zod must be a function that returns a zod validation function.`);
 }
 
 /**
@@ -416,11 +416,11 @@ export const validatePropZod = (prop, schema) => {
 export const validatePropDb = (prop, schema) => {
     if (['virtual'].includes(prop.type)) {
         if ('db' in prop)
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.db should not exist.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.db should not exist.`);
         return;
     }
     if (!isFunction(prop.db))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.db must be a function that returns a drizzle column specification.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.db must be a function that returns a drizzle column specification.`);
 }
 
 /**
@@ -433,13 +433,13 @@ export const validatePropDb = (prop, schema) => {
  */
 export const validatePropDefault = (prop, schema) => {
     if (['virtual'].includes(prop.type) && prop.default && !isString(prop.default))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.default must be a string.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.default must be a string.`);
     if (!prop.required)
         return;
     if (!('default' in prop))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.default must be set.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.default must be set.`);
     if (['string', 'id', 'json'].includes(prop.type) && !(isString(prop.default) || isFunction(prop.default) || isUndefined(prop.default)))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.default must be a string or a function or undefined.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.default must be a string or a function or undefined.`);
 }
 
 /**
@@ -453,13 +453,13 @@ export const validatePropDefault = (prop, schema) => {
 export const validatePropFilterQuery = (prop, schema) => {
     if (!['enum', 'string'].includes(prop.type)) {
         if ('filterQuery' in prop)
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.filterQuery should not exist.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.filterQuery should not exist.`);
         return;
     }
     if (!('filterQuery' in prop)) // not required
         return;
     if (!isBoolean(prop.filterQuery))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.filterQuery must be a boolean.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.filterQuery must be a boolean.`);
 }
 
 /**
@@ -473,13 +473,13 @@ export const validatePropFilterQuery = (prop, schema) => {
 export const validatePropFilterQualifier = (prop, schema) => {
     if (!['enum', 'string', 'int', 'boolean', 'id'].includes(prop.type)) {
         if ('filterQualifier' in prop)
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.filterQualifier should not exist.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.filterQualifier should not exist.`);
         return;
     }
     if (!('filterQualifier' in prop)) // not required
         return;
     if (!isBoolean(prop.filterQualifier))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.filterQualifier must be a boolean.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.filterQualifier must be a boolean.`);
 }
 
 /**
@@ -493,13 +493,13 @@ export const validatePropFilterQualifier = (prop, schema) => {
 export const validatePropFilterRange = (prop, schema) => {
     if (!['int', 'date'].includes(prop.type)) {
         if ('filterRange' in prop)
-            throw ving.ouch(442, `${schema.kind}.props.${prop.name}.filterRange should not exist.`);
+            throw ving.ouch(442, `${formatPropPath(prop, schema)}.filterRange should not exist.`);
         return;
     }
     if (!('filterRange' in prop)) // not required
         return;
     if (!isBoolean(prop.filterRange))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.filterRange must be a boolean.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.filterRange must be a boolean.`);
 }
 
 /**
@@ -512,7 +512,7 @@ export const validatePropFilterRange = (prop, schema) => {
  */
 export const validatePropRequired = (prop, schema) => {
     if (!isBoolean(prop.required))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.required must be a boolean.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.required must be a boolean.`);
 }
 
 /**
@@ -525,9 +525,9 @@ export const validatePropRequired = (prop, schema) => {
  */
 export const validatePropName = (prop, schema) => {
     if (!(/^[a-zA-Z]/.test(prop.name)))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.name must start with a letter.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.name must start with a letter.`);
     if (!(/[a-zA-Z0-9_]+/.test(prop.name)))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.name cannot contain any funky characters.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.name cannot contain any funky characters.`);
 }
 
 /**
@@ -541,9 +541,9 @@ export const validatePropName = (prop, schema) => {
 export const validatePropType = (prop, schema) => {
     const valid = ['boolean', 'int', 'date', 'enum', 'id', 'string', 'json', 'virtual']
     if (!valid.includes(prop.type))
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.type is invalid: ${prop.type}`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.type is invalid: ${prop.type}`);
     if (prop.type == 'virtual' && prop.relation.type != 'child')
-        throw ving.ouch(442, `${schema.kind}.props.${prop.name}.relation.type must be child when ${schema.kind}.props.type is virtual.`);
+        throw ving.ouch(442, `${formatPropPath(prop, schema)}.relation.type must be child when ${formatPropPath(prop, schema)}.type is virtual.`);
 }
 
 
@@ -631,4 +631,16 @@ export const validateOwner = (schema) => {
             throw ving.ouch(442, `Schema ${schema.kind} owner ${owner} doesn't match any known roles.`);
         }
     }
+}
+
+
+/**
+ * Outputs a prop name as an object path
+ * @param {object} prop The prop.
+ * @param {VingSchema} schema The schema for the prop
+ * @example
+ * formatPropPath(schema)
+ */
+export const formatPropPath = (prop, schema) => {
+    return `${schema.kind}.props.${prop.name}${prop.type == 'virtual' ? '{' + prop.relation.name + '}' : ''}`
 }